Refactor delegation components import/export
