+++ components/delegations/DelegatePicture.tsx
@@ -2,7 +2,7 @@ import { DelegateStatusEnum } from 'lib/delegates/constants';
 import { Box, Image } from 'theme-ui';
 import { Delegate } from 'types/delegate';
 
-export default function DelegatePicture({ delegate }: { delegate: Delegate }): React.ReactElement {
+export function DelegatePicture({ delegate }: { delegate: Delegate }): React.ReactElement {
   return (
     <Box sx={{ width: '41px', height: '41px', position: 'relative' }}>
       <Image
