+++ components/delegations/index.ts
@@ -1,0 +1,13 @@
+export * from './modals/Approval';
+export * from './modals/Confirm';
+export * from './modals/DelegateModal';
+export * from './modals/InputDelegateMkr';
+export * from './modals/TransactionInProgress';
+export * from './modals/TxDisplay';
+export * from './modals/TxFinal';
+export * from './modals/UndelegateModal';
+export * from './DelegateCard';
+export * from './DelegateContractExpiration';
+export * from './DelegateDetail';
+export * from './DelegateLastVoted';
+export * from './DelegatePicture';
